requirements.txt
# Bindings for the hermes protocol
pytoml
RPi.GPIO
smbus
paho.mqtt